requirements.txt
Flask==1.1.1
requests==2.22.0
python-dotenv==0.10.3
flask-restful==0.3.7